Added custom Deserializer for JSON (model)configurations
The use exploits injectable values conveying infos about the type of
configuration we are deserializing. The use of these values is similar
for xml and yaml deserialization and is the one adopted in
obtainConfiguration function in MappingUtils: in both cases the
deserializer returns a generic Configuration.

The Yaml beans were reorganized.

Minor bugfix in firewall serializer (both xml and yaml)