Added to CLI support to JSON on grpc interface
The original grpc objects defined by the previous group were reused.
Some functions from the original client (not the tosca one ) were 
reused and inserted in the ToscaClient which is the one used by the CLI.
The original Client was not deleted but is never used.
Modifications have been done on GrpcUtils to include ids in the Grpc to
Model conversion.

The feature was not deeply tested.